Split JsonElementBuilder into separate builders.
