+++ src/test/java/technology/sola/json/builder/JsonObjectBuilderTest.java
@@ -1,14 +1,14 @@
-package technology.sola.json;
+package technology.sola.json.builder;
 
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class JsonElementBuilderTest {
+class JsonObjectBuilderTest {
   @Test
-  void object() {
-    var result = JsonElementBuilder.startObject()
+  void test() {
+    var result = new JsonObjectBuilder()
       .addInt("int", 2)
       .addLong("long", 3L)
       .addFloat("float", 1.5f)
@@ -18,7 +18,7 @@ public class JsonElementBuilderTest {
       .addNull("null")
       .addArray("array", array -> array.addString("test"))
       .addObject("object", object -> object.addString("string", "value"))
-      .finish();
+      .build();
 
     assertEquals(2, result.getInt("int"));
     assertEquals(3L, result.getLong("long"));
@@ -29,30 +29,5 @@ public class JsonElementBuilderTest {
     assertTrue(result.isNull("null"));
     assertEquals("test", result.getArray("array").getString(0));
     assertEquals("value", result.getObject("object").getString("string"));
-  }
-
-  @Test
-  void array() {
-    var result = JsonElementBuilder.startArray()
-      .addInt(2)
-      .addLong(3L)
-      .addFloat(1.5f)
-      .addDouble(2.5)
-      .addBoolean(true)
-      .addString("value")
-      .addNull()
-      .addArray(array -> array.addString("test"))
-      .addObject(object -> object.addString("string", "value"))
-      .finish();
-
-    assertEquals(2, result.getInt(0));
-    assertEquals(3L, result.getLong(1));
-    assertEquals(1.5f, result.getFloat(2));
-    assertEquals(2.5, result.getDouble(3));
-    assertTrue(result.getBoolean(4));
-    assertEquals("value", result.getString(5));
-    assertTrue(result.isNull(6));
-    assertEquals("test", result.getArray(7).getString(0));
-    assertEquals("value", result.getObject(8).getString("string"));
   }
 }
